Revert "Revert "Merge remote-tracking branch 'remotes/origin/CalebFlick' into Programming""
This reverts commit c7bdf00ee3af94d7c4d40c4433e62ed5560002c8.

+++ Assets/Scripts/Combat/DamagePlayer.cs
@@ -9,7 +9,7 @@ using UnityEngine;
 public class DamagePlayer : MonoBehaviour
 {
     [SerializeField]
-    private float attackDamage = 10.0f;
+    private int attackDamage = 10;
 
     private void OnTriggerEnter(Collider other)
     {
